third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/declared/get-expected.txt

This is a testharness.js-based test.
PASS Getting a custom property not in the CSS rule returns null
PASS Getting a valid property not in the CSS rule returns null
PASS Getting a valid property from CSS rule returns the correct entry
FAIL Getting a valid custom property from CSS rule returns the correct entry assert_equals: expected "auto" but got " auto"
PASS Getting a list-valued property from CSS rule returns only the first value
PASS Declared StylePropertyMap.get is not case-sensitive
PASS Declared StylePropertyMap.get reflects changes in the CSS rule
Harness: the test ran to completion.
